Refactor javascript parser into modules
